requirements/base.txt
Django==2.1.2
django-authtools==1.6.0
django-extensions==2.0.7
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
six==1.11.0
typing==3.6.4
